Change: Move useUserTimezone hook to hooks directory
