+++ src/web/hooks/__tests__/useUserTimezone.jsx
@@ -7,7 +7,7 @@ import {describe, test, expect} from '@gsa/testing';
 
 import {setTimezone as setTimezoneAction} from 'web/store/usersettings/actions';
 
-import {rendererWith, fireEvent} from '../testing';
+import {rendererWith, fireEvent} from 'web/utils/testing';
 
 import useUserTimezone from '../useUserTimezone';
